+++ src/collective/impersonate/tests/test_setup.py
@@ -6,6 +6,13 @@ from plone import api
 import unittest
 
 
+try:
+    from Products.CMFPlone.utils import get_installer
+except ImportError:
+    # BBB for Plone 5.0 and lower.
+    get_installer = None
+
+
 class TestSetup(unittest.TestCase):
     """Test that collective.impersonate is properly installed."""
 
@@ -14,12 +21,20 @@ class TestSetup(unittest.TestCase):
     def setUp(self):
         """Custom shared utility setup for tests."""
         self.portal = self.layer['portal']
-        self.installer = api.portal.get_tool('portal_quickinstaller')
+        if get_installer is None:
+            self.installer = api.portal.get_tool('portal_quickinstaller')
+        else:
+            self.installer = get_installer(self.portal)
 
     def test_product_installed(self):
         """Test if collective.impersonate is installed."""
-        self.assertTrue(self.installer.isProductInstalled(
-            'collective.impersonate'))
+        if get_installer is None:
+            is_installed = self.installer.isProductInstalled(
+                'collective.impersonate')
+        else:
+            is_installed = self.installer.is_product_installed(
+                'collective.impersonate')
+        self.assertTrue(is_installed)
 
     def test_browserlayer(self):
         """Test that ICollectiveImpersonateLayer is registered."""
@@ -35,13 +50,22 @@ class TestUninstall(unittest.TestCase):
 
     def setUp(self):
         self.portal = self.layer['portal']
-        self.installer = api.portal.get_tool('portal_quickinstaller')
-        self.installer.uninstallProducts(['collective.impersonate'])
+        if get_installer is None:
+            self.installer = api.portal.get_tool('portal_quickinstaller')
+            self.installer.uninstallProducts(['collective.impersonate'])
+        else:
+            self.installer = get_installer(self.portal)
+            self.installer.uninstall_product('collective.impersonate')
 
     def test_product_uninstalled(self):
         """Test if collective.impersonate is cleanly uninstalled."""
-        self.assertFalse(self.installer.isProductInstalled(
-            'collective.impersonate'))
+        if get_installer is None:
+            is_installed = self.installer.isProductInstalled(
+                'collective.impersonate')
+        else:
+            is_installed = self.installer.is_product_installed(
+                'collective.impersonate')
+        self.assertFalse(is_installed)
 
     def test_browserlayer_removed(self):
         """Test that ICollectiveImpersonateLayer is removed."""